mcp/requirements.txt
mcp[cli]>=1.0.0
google-cloud-bigquery>=3.11.0
google-cloud-storage>=2.10.0
google-cloud-run>=0.10.0
google-cloud-logging>=3.5.0
